# nexora-backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Nexora API",
    "version": "1.0.0",
    "description": "Dedicated Virtual Account & Reconciliation Engine — Powered by Nomba. Provision customer-named NUBANs, reconcile inbound transfers automatically, and expose clean APIs for downstream integrations.",
    "contact": {
      "name": "Nexora Developer Support",
      "email": "dev@nexora.ng"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" },
    { "url": "https://api.nexora.ng", "description": "Production" }
  ],
  "security": [{ "BearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Your Nexora API key from Settings → API Credentials"
      }
    },
...
  },
  "paths": {
...
    "/webhooks/nomba": {
      "post": {
        "summary": "Nomba webhook receiver",
        "description": "Receives Nomba event notifications. Verifies HMAC-SHA256 signature using your webhook secret, deduplicates by `requestId`, and triggers the reconciliation engine. Always ACKs 200 immediately before processing. Supported events: `virtual_account.funded`, `transfer.success`, `transfer.failed`.",
        "tags": ["Webhooks"],
        "parameters": [
          { "in": "header", "name": "x-nomba-signature", "required": true, "schema": { "type": "string" }, "description": "HMAC-SHA256 hex digest of the raw request body" }
        ],
        "responses": {
          "200": { "description": "Acknowledged" },
          "401": { "description": "Invalid or missing signature" }
        }
      }
    }
  }
}